[Extensions] Regenerate tts externs
During the review of crrev.com/c/5253987 a comment in the API schema got
adjusted to use a reference to the associated type on the API, but the
externs didn't get re-generated. No functional change is expected from
this change (it just propagates a comment edit to the generated file),
but it fixes some pre-submit warnings that can show up when edited
adjacent files.

Bug: 41496696
Change-Id: I103dcebfc76c4173e8dfed91b2a355eb882d8aa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5536219
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300346}
NOKEYCHECK=True
GitOrigin-RevId: f20597031a83c368bfa123c8d03c856780083ae0