cc/canvas: Adds CorePaintFlags and uses it in two new ops
This does the following:
. Introduces CorePaintFlags, which contains the most commonly used
  state from PaintFlags.
. Adds two ops (draw line/draw arc) that use these ops.
. Increases max size of PaintOpBuffer that Canvas will use when
  Graphite is enabled.

I restricted increasing the PaintOpBuffer max size to graphite as
doing so for ganesh negatively impacted metrics. It was
necessary to increase the max size for graphite as not doing so
results in a regression.

How the minimal (lite) ops were introduced results in code
duplication. I plan on investigating strategies for reducing that
after this patch lands, as well as extending the usage of
CorePaintFlags to other PaintOps. As I suspect this will be rather
invasive, I'm hoping to land this first, and then do the cleanup.
I'm also hoping to get this into 126, which branches shortly.

Bug: 5524058

Change-Id: Ica1ea33aa074919fdea89754bdf840c5b2bda20a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5528095
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300625}
NOKEYCHECK=True
GitOrigin-RevId: 6ee403126fee84e363b7ff44e0fb5b7d2f7ccfe1