[field-diff] UpdatePropertySpecificDifferencesTextDecorationOrColor
This shifts UpdatePropertySpecificDifferencesTextDecorationOrColor
to the field-diff logic.

The one behaviour change in this patch is making TextEmphasisFill
trigger a full paint/layout invalidation.

This was/is hiding a near impossible bug to trigger. E.g. TextEmphasisFill triggers the use of different glyphs to be used
for text-emphasis. Almost always this will be the same size.

However it is *possible* for a font to have different sized glyphs for
this, hence trigger a full layout/paint invalidation.

There should be no other behaviour change.

Bug: 339300436
Change-Id: I37bd548b3773061a1b739bd68f878c1a734ae4a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5510986
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300844}
NOKEYCHECK=True
GitOrigin-RevId: b9d14756a1f454ec124f0c271542db1d62ca69aa