cc: Fix render surface dependency tracking for view transition.
We iterate through the list of layers in paint order to populate the
list of render surfaces drawn in a frame. This list must be in
dependency order, the target surface a surface draws into must be added
before it.

This is tricky for ViewTransition where the surface producing the
snapshot is not drawn into its target, but into the surface drawing the
ViewTransitionContentLayer which renders this snapshot. This is
currently handled by adding all surfaces for view transition layers
*before* iterating through the list of layers. The assumption being that
VT layers don't have contributions from other VT layers so they can be
iterated in any order. And adding their render surfaces first ensures
that the surfaces for layers generating their snapshots are correctly
added after them.

This assumption is incorrect for subframes since the VT layer for the
subframe contributes to the render surface which produces a snapshot
rendered by the VT layer for the main frame.

This patches fixes the above by adding `view_transition_target_id` on
cc::EffectNode. For each effect node which is producing a snapshot, the
new target id tracks which effect node this snapshot will eventually be
rendered into via its ViewTransitionContentLayer. We use this target id
to correctly order render surfaces.

R=vmpstr@chromium.org

Bug: 330917526
Change-Id: Id621676865a388135415a2199fb84ee0f9425242
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5539412
Owners-Override: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300849}
NOKEYCHECK=True
GitOrigin-RevId: 582fd6a4f4a0481fb674dc44b79921dd327506cb