TaskAttribution: Only propagate state when necessary
Now that SoftNavigationContext is being propagated, we no longer need to
propagate task state unless there is SoftNavigationContext to propagate.

This main complication is testing.  To make tests work, we need make
sure CPED is cleared at the end of the current task, which won't happen
automatically if there isn't a TaskScope on the stack -- which might not
be the case when using the test-only scheduler.taskId. To do this, this
CL exposes the existing ExecuteAfterCurrentTask() API via
MainThreadScheduler, guarded by a pass-key. Note that it isn't
sufficient to use a microtask since CPED is restored after a microtask
checkpoint.

Bug: 40942324
Change-Id: I07a69d1d090a6813447fae63d05bda3d30c4fd57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5347178
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301380}
NOKEYCHECK=True
GitOrigin-RevId: 1a763cf78bab8e3b6025238aeaf49bb0a4674b35