IDB: Add debug code for failure to lock Level DB LOCK file.
Not many features use this indirection layer (FilesystemImpl).
Only LevelDB (via ChromiumEnv) and the unzip component actively
use it. There is glue code for using SQLite in a sandboxed
process, `SandboxedVfsWrapper`, which uses it, but we don't
actually run any SQLite DBs in a sandboxed process. The unzip
component doesn't try to lock files AFAICT. Therefore this change
should only be relevant for LDB.

The debug dump is inserted at the line which would indicate that
the lock is being held by the same Chromium process already, which
indicates a programming error somewhere in Chromium/IDB or LDB.
Luckily IDB's LDB is not in a sandboxed process, so the stack should
be enlightening if it's triggered. If we don't get any dumps from
this line, it will suggest that acquisition of the lock is failing
at the OS API level, i.e. that a different process is holding the
lock. We can't do much about that except perhaps ignore the lock
and forge ahead anyway.

Bug: 340398745
Change-Id: I2b15d4cca330f3b5405d0c5eed6a9b082065c602
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5538989
Reviewed-by: Austin Sullivan <asully@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301491}
NOKEYCHECK=True
GitOrigin-RevId: 4490965817518382bc7af3b808b69e84e29a26d5