[IntersectionObservation] Fix needs_occlusion_tracking bookkeeping
This fixes three related bugs in occlusion tracking:

1. When an IntersectionObserver using trackVisibility is created in an
OOPIF, there is some latency in notifying the parent renderer that
occlusion tracking is required for the iframe. During that interval,
the parent renderer might send a 'guaranteed not occluded' message to
the child, causing the IntersectionObserver to report a false
positive. With this change, we will never tell an OOPIF that it's
"guaranteed not occluded" if the OOPIF hasn't requested occlusion
information. This is fixed in frame_view.cc.

2. When an OOPIF runs LocalFrameView::RunIntersectionObserverSteps,
it might get a transient `false` value when recomputing
needs_occlusion_tracking, due to an early return from
UpdateViewportIntersectionsForSubtree (which has a couple of
early-return edge cases). When that happens, we should *not* relay
the `false` value back to the parent process, because it can create
a temporary gap in occlusion reporting from the parent. We should
instead preserve the previous value until it can be successfully
recomputed on a subsequent BeginMainFrame. This is fixed in
local_frame_view.cc.

2. For OOPIF's, needs_occlusion_tracking is stored in two places:
on RemoteFrameOwner::needs_occlusion_tracking_ in the iframe process,
and in RemoteFrameView::needs_occlusion_tracking_ in the parent
process. When the first one changes in the iframe process, it IPC's
to the parent process to synchronize the second one. To keep them
consistent, there should be no other point of entry into
RemoteFrameView::SetNeedsOcclusionTracking. Currently, that method is
also called from RemoteFrameView::Dispose, but that call is unnecessary
and it can cause the two values to get out of sync. This is fixed in
remote_frame_view.cc.

Bug: chromium:333708039
Change-Id: Id1dd08eefb5aa201651b2c6950ea61f3b4a081a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5540331
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301762}
NOKEYCHECK=True
GitOrigin-RevId: 78bd1f0eef2612f22dac87d0335e8714a0ecee89