RubyLB: LineBreaker should have a RubyBreakTokenData
There was code to extract a RubyBreakTokenData from the InlineBreakToken
passed to the constructor and use it in BreakLine(), but it did not
handle the case where a RubyBreakTokenData needs to be transmitted when
calling NextLine() multiple times with a single LineBreaker instance.
So, this CL adds LineBreaker::ruby_break_token_.

InlineItemResultRubyColumn also has the ruby_break_token_ field in
order to pass a RubyBreakTokenData from HandleRuby() to NextLine().

Bug: 324111880
Change-Id: If18ad5d81dfb5ef589becc668edd5f3c87f57055
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5541295
Auto-Submit: Kent Tamura <tkent@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301804}
NOKEYCHECK=True
GitOrigin-RevId: 942362472abbde5999d79ec7069be5d7c201edef