[css-shadow-parts] Disallow tag names under a restricting pseudo element
We currently have a bug where e.g. ::part(x):is(div) can match,
even though we're not supposed to reveal that kind of information
on the part. We do check whether or not a given simple selector is
valid following certain pseudo-elements, however this check is only
carried out for selectors that use ConsumeSimpleSelector, which is
not the case for tag names.

This CL fixes the issue by failing the parse if we see any tag name
selector under any restricting pseudo element.

Fixed: 41491988
Change-Id: I984a0a755455a8d1030bba59cc5aac472f6b9339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5542441
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301833}
NOKEYCHECK=True
GitOrigin-RevId: e3eb2ff0dff4cd0c00a54a0a8dfc0b6f5327e81d