Fix thread safety analysis annotations on base::AutoTryLock
Prior to this CL, base::AutoTryLock's constructor was annotated with
EXCLUSIVE_LOCK_FUNCTION, a Thread Safety Analysis [1] annotation that
indicates that the function infallibly acquires the lock. In reality, it
just *tries* to acquire the lock, so potentially-unguarded access was
flying under the radar.

It appears it's structurally impossible to annotate a constructor as a
fallible locking operation. The EXCLUSIVE_TRYLOCK_FUNCTION(...)
annotation exists, but it requires a boolean parameter corresponding to
the function's return type. Constructors have no return type, so it
seems this annotation simply cannot be applied to constructors. This
problem is described in a newly-filed LLVM issue [2].

As a workaround, this CL annotates `bool AutoLock::is_acquired()` with
EXCLUSIVE_TRYLOCK_FUNCTION(true) instead of the constructor. The
accessor does not really acquire the lock, but that's an implementation
detail from the caller's perspective. The effect is that the analysis
now requires callers to check is_acquired() before using guarded data.

[1]: https://clang.llvm.org/docs/ThreadSafetyAnalysis.html
[2]: https://github.com/llvm/llvm-project/issues/92408

Bug: 340196356
Change-Id: I98a37e9c9260eee417250190b770a2e2d8cee720
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5542379
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301979}
NOKEYCHECK=True
GitOrigin-RevId: f97ab60f89600edf9068e5a7a87c6413009de5ed