[UnblockTouchMoveEarlier] Avoid blocking second touchmove
When the feature is enabled, MainThreadEventQueue unblock queued
blocking touchmoves after the touchstart and the first touchmove are
handled and they didn't call preventDefault(), instead of waiting
for the browser to send TouchScrollStarted [1].

This avoids the racing condition between:
- The dispatching of the second queued blocking touchmove, and
- TouchScrollStarted after the roundtrip from the renderer to the
  browser
when the touchstart handler and/or the first touchmove handler are
slow, thus avoids the unnecessary touchmove handler from blocking
scroll while the scroll has already started.

[1] The logic to unblock queued touchmoves on TouchScrollStarted is
still kept because it may address the issue in different scenarios.

Bug: 339015752
Change-Id: I7160b77011fc1b1bc80579bdd5f93118b8bad2fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5527265
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302082}
NOKEYCHECK=True
GitOrigin-RevId: 51756a8fbed1fb70fe148baecb9448cbee05237a