Enable crabbyavif build and feature behind a flag
The CFI and msan issues are now fixed/worked around. This CL
enables building crabbyavif and exposes it behind a runtime flag.

This CL still does not enable the fuzzer as that can be done in a
follow-up.

CFI: https://crrev.com/c/5536152
Msan: https://crrev.com/c/5542623

Binary-Size: Part of the size increase is temporary as we roll out CrabbyAvif. When libavif is removed, some of the size will be recovered.
Bug: b/323754052
Change-Id: I4ef994929c015a364bb4c81114a1dcffe82467aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5542730
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Wan-Teh Chang <wtc@google.com>
Commit-Queue: Vignesh Venkat <vigneshv@google.com>
Cr-Commit-Position: refs/heads/main@{#1302114}
NOKEYCHECK=True
GitOrigin-RevId: b1f2281f75c4ffb28a5ae35ffb050dfeec1f4b19