Migrate from CTFontCopyVariationAxes to CTFontCopyVariationAxesInternal
Mac's SPI `CTFontCopyVariationAxesInternal()` is working faster than a
public `CTFontCopyVariationAxes()` because it does not localize
variation axis name string, see [0]. Since we don't need localized
variation axis name, we are replacing `CTFontCopyVariationAxes()` calls
with `CTFontCopyVariationAxesInternal()`. We are doing that only for
MacOS 13+ because these are our benchmarking platforms and because
`CTFontCopyVariationAxesInternal()` is only supported on MacOS 12+.

Request for public API in FB13788219.

[0] https://github.com/WebKit/WebKit/commit/1842365d413ed87868e7d33d4fad1691fa3a8129

Change-Id: I33af999f99fff9990252ce69514208c6d31e627d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5538014
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Munira Tursunova <moonira@google.com>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302213}
NOKEYCHECK=True
GitOrigin-RevId: cf9006d779b93fc4ce00213949db1e4d06def45f