[A11y] Process update only when options bounds are ready
When there are more than 1000 options in a select, the first 1000 are
included in the first layout and the rest are included later. However,
only the bounds of the options from the first layout are sent to the
accessibility tree, causing an index out of bounds error when trying to
access the bounds of the 1001st option element.

This change sends updated bounds to the tree and also prevents the
processing of the bounds before all of them are ready.

Fixed: 338646632
Change-Id: I2e6e8eaf473e3030aef9e6425201fba3672f733f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5532661
Commit-Queue: Jocelyn Tran <jocelyntran@google.com>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302226}
NOKEYCHECK=True
GitOrigin-RevId: 7b5a09f8d44399bdf4573f5194752ff0c7caa71f