Trigger LCP for SimplifyLoadingTransparentPlaceholderImage
When the "SimplifyLoadingTransparentPlaceholderImage" feature was
enabled, LCP was not previously triggered for transparent placeholder
images as we do not decode the image, and hence didn't have the size of
the painted image. If the "ExcludeLowEntropyImagesFromLCP" feature is
enabled, the bpp of an image must be above a threshold to be considered
for LCP. This means that for the transparent placeholder images to be
included in LCP calculations, they need to have a content size for
entropy as well. This CL sets this value.

Bug: 336681970
Change-Id: Id3aaad745ad863395d9c2fd06c7c90d923bba6b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5528344
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302363}
NOKEYCHECK=True
GitOrigin-RevId: 5c410c359ad43a320b1ec16e5455da1a78244cd9