[text-box-trim] Test when the line is taller than strut
As the `text-box-trim` trims at the specified metrics of the
line box, objects taller than that are trimmed, by overflowing
to over and under side of the line box.

This test verifies only the block-start side, but can be
extended to the block-end side when it's implemented.

Bug: 40254880
Change-Id: I7be386cff8f46f58b164c01381253d1b2cbe034f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5547519
Reviewed-by: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302464}
NOKEYCHECK=True
GitOrigin-RevId: 276a638a08331ea155cef25c5c08d6696cdc2e93