paint: remove dead code in HighlightPainter
Since CL:5019716 and CL:5334286, the PaintDecoratedText method added
in CL:4801206 is only used for composition marker foregrounds.

This patch removes the unused parameters, and renames that method (and
the equally-confusingly-named Paint) to make their purpose clearer.

No test changes because this is a refactor.

Bug: 339298411
Change-Id: I4982224d7dac084d5d663a96837a1364f7cd19cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5539143
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Delan Azabani <dazabani@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1302707}
NOKEYCHECK=True
GitOrigin-RevId: 7e375ac65aa8841685443c6742c2242415b278bd