Add a texture parameter to `endWebGPUAccess`.
This CL adds the GPUTexture to our IDL, and updates all tests to
conform to the new API. At present, the passed-in texture is
checked to verify that it's not destroyed, but we don't fully
validate it and don't use it to restore the canvas contents. In
the spirit of small CLs, that work will be done in a followup.

Bug: 339846593
Change-Id: Ica79a9c3e21a5385b341019cecb1479297a43bb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5538688
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Cr-Commit-Position: refs/heads/main@{#1302741}
NOKEYCHECK=True
GitOrigin-RevId: eb97ea8d70a20a41d26d6b8b081a80198c7fd13c