[Interop] Don't use aria-labelledby result if it's only whitespace or empty
This subtest is failing because some of the whitespace in the
aria-labelledby traversal is non-collapsed and shows up in the result,
which prevents the aria-label from being applied.

Fix this by only using the aria-labelledby result only if it is
non-empty and non-whitespace-only.

Note, the spec [1] does not explicitly state that whitespace should
be ignored here. However, other engines do ignore the whitespace; see
test results at [2]. And the spec *does* say that whitespace should
be ignored in aria-label [3]. So I think there's a good chance the
spec will be changed for this.

I filed a spec issue on this at [4]. That may take some time to
resolve, so in the meantime I think the right thing to do is to align
with the behavior of other browsers; hence this CL.

[1] https://w3c.github.io/accname/#comp_labelledby
[2] https://wpt.fyi/results/accname/name/comp_label.html?label=experimental&label=master&aligned&view=interop&q=label%3Ainterop-2024-accessibility
[3] https://w3c.github.io/accname/#comp_label
[4] https://github.com/w3c/accname/issues/233

Bug: 338348441
Change-Id: Iffb623023e5ee3b903f2c243ae985efa5179135e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5540421
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302913}
NOKEYCHECK=True
GitOrigin-RevId: f9ad6c790b53d8e615897fedd231389a7fc5e2ae