Update Sec-CH-Width header for lazy loaded images
The Sec-CH-Width request header is set when the lazy loaded image
request is initially created. Lazy loaded requests are deferred so the
actual request is not made until later. When the actual request is made,
the resource width may be different, and this patch updates the resource
width header.

Fixed: 335630145
Change-Id: I1316ef3519b8e62a5a932a212292c30556cb9479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5548501
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1303180}
NOKEYCHECK=True
GitOrigin-RevId: 8a776d7a77ae56f96653cb6e324fc54bcc615bc0