Backlight OCR: Various fixes to the StitchChildTree action
1. The default target role should be kUnknown not kNone.
kNone is used for presentational objects (see ARIA Spec).
2. Minus one was specified for the default value of
`AXActionData::target_node`
but in the code we check for kInvalidAXNodeId.
3. We were verifying in Blink if the target node ID in
`AXActionData` was equal to AXID (the type)
instead of the AXObjectID() which returns the ID of the current
AXObject.
4. Updated Mojom traids to reflect the use of kUnknown
and -1 for the target_role and target_node_id respectively.

AX-Relnotes: n/a.
Change-Id: I67bcdaa142b5b694f1e779643930dd23e11d1e96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5544844
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Commit-Queue: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1303389}
NOKEYCHECK=True
GitOrigin-RevId: b3ed43812659c2d9e501fce9ba1f6170223d0ebd