Don't create CSSNumericValue objects for lengths with calc-size().
Similar to the change made for anchor functions in
https://crrev.com/0d6c5b1397b718a4deffaf3f1b8c65dc125fba61 this avoids
trying to construct typed OM objects for a case (the sizing keywords in
calc-size()) where we would fail to do so (and crash).  This does so for
all calc-size() values rather than only those with sizing keywords
because this is the main use case for calc-size() and the change is
easier that way.

Bug: 313072
Change-Id: I9ad3d992836a9fe6eca277b66ed317cf6023b518
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5555160
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1303887}
NOKEYCHECK=True
GitOrigin-RevId: 3ae603ad4984fa277a826fbf4b7d85a254c13a9b