Roll unrar to v7.0.8
Updates unrar to v7.0.8 as extracted from
https://www.rarlab.com/rar/unrarsrc-7.0.8.tar.gz and mirrored at
https://github.com/aawc/unrar/releases/tag/v7.0.8.

Upstream release notes: https://www.rarlab.com/unrar7notes.htm

The first patchset is a snapshot of the source. The Chromium patches and
readme updates are then applied on top of that in subsequent patchsets.

This patches out SSE support by disabling the defining of the USE_SSE
flag. This was already patched out of the Chromium build in previous
versions on Windows; this CL patches it out for all platforms. This
optimization is not needed, and would be problematic for the build
because it uses function multiversioning which introduces static
initializers.

Of note, I introduced a macro UNRAR_FATAL_BAD_ALLOC for the memory
errors that either call base::TerminateBecauseOutOfMemory() (for
UNRAR_NO_EXCEPTIONS) or throw std::bad_alloc() in certain files.
There is no functional change; it just makes the diff nicer.

Bug: 336357535
Change-Id: I86b751dbc3a9632a933b0c16234d3fd6f2304cfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5544995
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1304132}
NOKEYCHECK=True
GitOrigin-RevId: cfed5c1758b4cd795e5c48608aee9fc128230c0f