Fix a bug in GenericFontFamilySettings copy constructor.
This is a follow-up of Id219621c8cc1513c6a0d124f06522f8a0044269c.
The copy constructor of GenericFontFamilySettings did not copy the
first_available_font_for_families_ map.
This caused a crash when the copy of GenericFontFamilySettings was used
to create a FontDescription.

Bug: 327545434
Change-Id: Ifa495675a9a4e6f1f2409c70ea1bf7f86dde7602
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5554151
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1304162}
NOKEYCHECK=True
GitOrigin-RevId: 17bdb170168b15c2513ece9a3326e407fc5219d6