[text-box-trim] Test block-end with Ruby
This patch adds a test of the `text-box-trim` property for the
block-end behavior when a Ruby hangs to the block-end side.

The test and reference for the block-start added in
crrev.com/c/5539108 are updated to share the reference file.

Bug: 40254880
Change-Id: Ie068940eff74aab2eb4baa7588b37baaef59c590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5553008
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1304323}
NOKEYCHECK=True
GitOrigin-RevId: 52793450f7e4f84619dff3aac2ae7568f6c6a7ea