IDB: add more logging for locking failures.
The intention of IndexedDB.BackingStoreCloseDuration and
IndexedDB.ContextShutdownDuration is to determine how often shutdown
of prior databases may interfere with subsequently opened ones. (See
bug for details.)

The intention of `LevelDBEnv.LockFileInUseByThisProcess` is to
disambiguate the sources of the FILE_IN_USE error condition.

The intention of `LevelDBEnv.LockFileSuccessAttempts` is to get a feel
for how long retrying might actually be useful.

Bug: 340398745
Change-Id: I0c42ba8cdf521137e4e533fe9a3321bdaaeec7be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5555885
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1304478}
NOKEYCHECK=True
GitOrigin-RevId: 91d6e6658a206e5fab3ff1dfd0443aa163483e57