Make CompressionDictionaryTransportV2 Origin Trial name to test only
The Origin Trial of CompressionDictionaryTransportV2 is scheduled to end
in M125. The M125 bunch was already cut. So this cl changes the
`origin_trial_feature_name` to make the existing OT tokens invalid for
M126 and later.

I forgot to land this CL in M126. So, I'd like to merge it to the M126
branch.

Bug: 1413922, 342247578
Change-Id: I2ad8e977ccc8a28fc37d7b785295735f19c384c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5492561
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1304882}
NOKEYCHECK=True
GitOrigin-RevId: e0d8992dc6e36e77575f07eefb6b855d05ad7acc