Fix bug in positioning the media element overflow menu
Since anchor names are tree-scoped references, if they're defined in
the UA stylesheet, they're global to the page. That means that if
there are multiple media elements on a page, and all of them define
an `anchor-name`, then any top-layer elements trying to anchor to that
anchor-name will "pick" the last one on the page. That means that
clicking any of the overflow menu buttons results in an overflow menu
that is anchored to the *last* such element.

The change is simply to rely on the new `anchor-scope` property to
scope all anchor names beneath the media controls element. This
also keeps the styles in the UA stylesheet, which means this usage
of anchor positioning should not be use-counted.

Bug: 40281992,334817145
Fixed: 341741271
Change-Id: I9e93b1e3d7e8a5792d4ff9f553a885b6db28ebbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5547787
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1304912}
NOKEYCHECK=True
GitOrigin-RevId: f235aed1f0b36e760cd1968e8077927d69cf3bd4